move containers into folder
