+++ test/cpp/api/any.cpp
@@ -1,8 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <torch/nn/modules/any.h>
-#include <torch/nn/modules/linear.h>
-#include <torch/utils.h>
+#include <torch/torch.h>
 
 #include <test/cpp/api/support.h>
