+++ tools/build_variables.py
@@ -191,12 +191,12 @@ def add_torch_libs():
         "torch/csrc/api/src/nn/modules/dropout.cpp",
         "torch/csrc/api/src/nn/modules/sparse.cpp",
         "torch/csrc/api/src/nn/modules/fold.cpp",
-        "torch/csrc/api/src/nn/modules/functional.cpp",
         "torch/csrc/api/src/nn/modules/linear.cpp",
         "torch/csrc/api/src/nn/modules/loss.cpp",
-        "torch/csrc/api/src/nn/modules/named_any.cpp",
         "torch/csrc/api/src/nn/modules/pooling.cpp",
         "torch/csrc/api/src/nn/modules/rnn.cpp",
+        "torch/csrc/api/src/nn/modules/container/functional.cpp",
+        "torch/csrc/api/src/nn/modules/container/named_any.cpp",
         "torch/csrc/api/src/optim/adagrad.cpp",
         "torch/csrc/api/src/optim/adam.cpp",
         "torch/csrc/api/src/optim/lbfgs.cpp",
