+++ torch/csrc/api/include/torch/nn/modules.h
@@ -1,16 +1,19 @@
 #pragma once
 
-#include <torch/nn/modules/any.h>
+// Containers
+#include <torch/nn/modules/container/any.h>
+#include <torch/nn/modules/container/functional.h>
+#include <torch/nn/modules/container/modulelist.h>
+#include <torch/nn/modules/container/named_any.h>
+#include <torch/nn/modules/container/sequential.h>
+
+// Layers
 #include <torch/nn/modules/batchnorm.h>
 #include <torch/nn/modules/conv.h>
 #include <torch/nn/modules/dropout.h>
 #include <torch/nn/modules/sparse.h>
 #include <torch/nn/modules/fold.h>
-#include <torch/nn/modules/functional.h>
 #include <torch/nn/modules/linear.h>
 #include <torch/nn/modules/loss.h>
-#include <torch/nn/modules/modulelist.h>
-#include <torch/nn/modules/named_any.h>
 #include <torch/nn/modules/pooling.h>
 #include <torch/nn/modules/rnn.h>
-#include <torch/nn/modules/sequential.h>
